feat:edit and delete post,comment,reply done

+++ src/applications/use-cases/comment/deleteComment.usecase.ts
@@ -5,8 +5,9 @@ export class DeleteCommentUsecase {
   constructor(
     private commentRepository: CommentRepository
   ) {}
-  async execute(commentId: string) {
+  async execute(commentId: string, userId: string) {
     await this.commentRepository.verifyCommentAvailability(commentId);
+    await this.commentRepository.verifyCommentOwnership(userId, commentId);
     await this.commentRepository.deleteComment(commentId);
   }
 }
