+++ src/applications/use-cases/posts/deleteMedia.usecase.ts
@@ -4,8 +4,9 @@ export class DeleteMediaUsecase {
   constructor(
     private postRepository: PostRepository
   ) {}
-  async execute(postId: string) {
+  async execute(postId: string, userId: string) {
     await this.postRepository.verifyPostAvailability(postId);
+    await this.postRepository.verifyPostOwnership(userId, postId);
     await this.postRepository.deletePost(postId);
   }
 }
