+++ src/domains/repositories/comment/comment.repository.ts
@@ -1,9 +1,12 @@
 import { CommentM } from "src/domains/model/comment";
+import { EditCommentDto } from "src/presentations/comment/dto/editComment.dto";
 
 export interface CommentRepository {
   createComment(comment: CommentM ): Promise<void>
   verifyCommentAvailability(commentId: string): Promise<boolean>
+  verifyCommentOwnership(userId: string, commentId: string): Promise<boolean>
   getCommentsByPostId(postId: string): Promise<CommentM[]>
   getCommentById(commentId: string): Promise<CommentM>
+  editComment(commentId: string,comment: Partial<EditCommentDto>): Promise<void>
   deleteComment(commentId: string): Promise<void>
 }
