+++ src/domains/repositories/post/post.repository.ts
@@ -1,10 +1,13 @@
 import { PostM } from "src/domains/model/post";
-import { PostMediaDto } from "src/presentations/posts/dto/postMedia.dto";
+import { EditMediaDto } from "src/presentations/posts/dto/editMedia.dto";
 
 export interface PostRepository {
   createPost(post: PostM): Promise<void>
   verifyPostAvailability(id: string): Promise<boolean>
+  verifyPostOwnership(userId: string, postId: string): Promise<boolean>
+  getPaginatedPosts(page: number, limit: number): Promise<{ posts: PostM[]; total: number }>
   getPostById(id: string): Promise<PostM>
   getDetailedPostById(id: string): Promise<PostM>
+  editPost(id: string, post: Partial<EditMediaDto>): Promise<void>
   deletePost(id: string): Promise<void>
 }
